Update oxygen_pre_deployment.md
What about taking water samples during the recovery? The ideal is to take at the beginning and at the end of the mission? (Line 123)
During the 8 hours, the temperature of the lab can change if there is no use of AC, so turning the AC could cause you to have a big range of temperature especially if you leave the sensor during the night! (Line 128).