Typos corrected
Please, feel free to accept/reject them :)
I remove reference to PLOCAN document since we haven't received the protocol yet so we don't have a reference for this. 
Soeren and Tom: do you think we should move sections 4.5.1.2 Coomunicating with the sensor using a terminal and 4.5.1.3. Calibration procedure to an annex? These are very specific for aanderaa optodes, so don't know what's your oppinion on this.