Typos corrected
Feel free to reject/accept them :)
At the end of section 6.1 there is a reference to a paragraph but no link, do we need to remove it? Or there is a document we can reference to?
I deleted reference to Fig 7.1 at the end of the last section, I think the diurnal cycle is not related with it?