add callums contribution
Thanks @callumrollo for your support in maintaining the SOPs. This is key for them to fly long term. 

Please adapt your contributions. 

Note that we are writing a summary paper for Frontiers. @tomhull @patricialg are leading this work and are happy to have fresh support!
Before we submit the Frontiers paper we have to ensure to make a new release v1.1.X or so submit it to OBPS to ensure to have you as co-author!